requirements.txt
# Core pipeline dependencies
dvc>=3.0.0
numpy>=1.24
soundfile>=0.12
scipy>=1.10
pyyaml>=6.0

# Phoneme recognition model
torch>=2.0
transformers>=4.38
datasets>=2.18

# Plotting
matplotlib>=3.7

# Data download (optional – only for setup_data.py)
# huggingface_hub>=0.20
